Changed admission by age to use proper csv generator

+++ CoronavirusInTheUKDashboard.Api.Service.Models/Models/Records/Ages/AdmissionsByAgeItemRecord.cs
@@ -8,7 +8,7 @@ namespace CoronavirusInTheUKDashboard.Api.Service.Models.Models.Records.Ages
     {
         public string AgeRange { get; set; }
         public double Rate { get; set; }
-        public double Cumulative { get; set; }
-        public double New { get; set; }
+        public long Cumulative { get; set; }
+        public long New { get; set; }
     }
 }
